+++ pyam/utils.py
@@ -1,4 +1,3 @@
-import copy
 import os
 import itertools
 import string
@@ -18,7 +17,6 @@ try:
 except ImportError:
     pass
 
-from pyam.core import IamDataFrame
 from pyam.logger import logger
 
 # common indicies
@@ -369,19 +367,3 @@ def to_int(x, index=False):
         return x
     else:
         return _x
-
-
-def concat(dfs):
-    """Concatenate a series of dataframes together"""
-    if not hasattr(dfs, '__iter__'):
-        raise ValueError('Input data must be iterable')
-
-    df = None
-    for _df in dfs:
-        if not isinstance(_df, pyam.IamDataFrame):
-            raise TypeError('Input contains non-dataframe')
-        if df:
-            df.append(_df, inplace=True)
-        else:
-            df = copy.deepcopy(_df)
-    return df
